move categoryPickerFragment to separate directory
